requirements.txt
# Core Dependencies
langchain_openai==0.0.2
python-dotenv==1.0.0
rich==13.7.0
pydantic==2.5.2

# Document Processing
python-docx==0.8.11
python-pptx==0.6.21
Pillow==10.1.0
reportlab==4.0.8

# Data Processing
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
openpyxl==3.1.2

# Web and Networking
aiohttp==3.9.1
requests==2.31.0
websockets==12.0

# File Operations
pathlib==1.0.1
pickle-mixin==1.0.2

# Utilities
tqdm==4.66.1
colorama==0.4.6
psutil==5.9.6

# Google API
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0

# Microsoft Office Integration
office365-rest-python-client==2.4.1

# Cloud Storage
dropbox==11.36.2
boto3==1.33.6

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Testing
pytest==7.4.3
pytest-asyncio==0.23.2
pytest-cov==4.1.0